Added first part of test1.
Test one starts by initializing a priority queue with a for loop through a set of integers. It then moves onto intitializing a std::vector of std::string then uses a ranged based for loop to iterate over the vector and push the strings onto the priority queue. Then use template display function to std::cout the contents of the priority queue's using pass b value so nothing is changed, as const correctness wont work with display as we need to pop the variable of the queue.